refactor: RestaurantSyncService 를 4개의 하위클래스 및 파사드패턴으로 분리
